refactor: split task recovery ownership
